Better transfer options and on call transfer support
Added transfer menu for calls (in contact plugins that support the transfer menu).  Added option for contact plugins to provide a default string (ie @yourdomain.com) to transfer numbers if no "@" is found in the transfer url.  These changes to the contact plugin system are breaking, and require some functions to be renamed.  You can just rename the old functions to the new names but it is best to return the MenuItems directly now rather than adding them to the context menu itself.
Number stripping options for plugins to be able to strip numbers off the call log.
Better handling for Jabra headsets as sometimes they change names and lose their reference.